Production level logging for the plugin. (#137)
We print logs through Facebook logger only. This way application could customize if logs need to be saved on disk or have some custom processing.
Instead of printing stacktraces we let FLog class to figure out how to handle it.
Log levels are changed in the way that important logs wouldn't be filtered out by default log policy in release.