fix: fixed the `source` field (#396)
The `source` field was causing errors when attempting to link the library via CocoaPods in iOS. It was unable to resolve the source since the tagging strategy seems to have changed to not include the "v" in the tag name and also, the URL needed to have the "git+" removed from it. I did not want to update the `package.json` file as well, so I just added in the source URL into the podspec directly.