Add prominent instruction for enabling Promises to Readme (#446)
As there is no mention of `SQLite.enablePromise(true);` in the Readme, I expected promises to work out of the box. 
After some time verifying that callbacks work but no result is returned from the functions that are supposed to return promises, I was about to file an issue, at which point I found #409 . 
I think adding this to the Readme, prominently at the top, would be a quick solution that prevents this happening to anyone else.